libvirt: avoid changing UUID when redefining nwfilters
libvirt >= 1.2.7 enforces that when you re-define a network
filter you can't change the UUID. ie name + uuid must match.
Since Nova was not including any UUID in the XML it sent, it
would always get a random UUID generated, which would cause
failures when re-defining an existing filter. The result
was that Nova would fail to start up and fail to migrate
if there was an existing guest running. The fix is to query
libvirt to see if the nwfilter already exists, and extract
the UUID from its XML and use that when re-defining it.

(cherry picked from commit 2418a9dce86280fbf269f68f7d7c9b739f662ef1)

Cherry-pick addapted to Folsom code --BranoZ