VMware: Refactor the image transfer
The image transfer is unnecessary complicated and buggy. When
transferring streamOptimized images we have to update the progress of
the NFC lease to prevent timeouts.
This patch replaces the complex usage of blocking queues and threads with
a simple read+write loop. It has the same performance and the code is
much cleaner. The NFC lease is updated with the loopingcall utility.

Closes-Bug: #1546454
Closes-Bug: #1278690
Related-Bug: #1495429
Change-Id: I96e8e0682bcc642a2a5c4b7d2851812bef60d2ff
(cherry picked from commit 2df83abaa0a5c828421fc38602cc1e5145b46ff4)